Split tuples' tests
